pyproject.toml
[tool.pysen]
version = "0.10"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 88
py_version = "py311"
mypy_ignore_packages = ["amaterus_admin_gradio.graphql_client.*"]

  [[tool.pysen.lint.mypy_targets]]
    paths = ["."]

[tool.pysen.lint.source]
  excludes = ["amaterus_admin_gradio/graphql_client/"]

[tool.ariadne-codegen]
queries_path = "queries/"
schema_path  = "schema.graphql"
target_package_path = "amaterus_admin_gradio/"
async_client = false

[tool.poetry]
name = "amaterus-admin-gradio"
version = "0.1.0"
description = ""
authors = ["aoirint <aoirint@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
gradio = "^3.44.3"
requests = "^2.31.0"
pydantic = "^2.3.0"
python-dotenv = "^1.0.0"
beautifulsoup4 = "^4.12.2"
html5lib = "^1.1"
ariadne-codegen = {extras = ["subscriptions"], version = "^0.9.0"}


[tool.poetry.group.dev.dependencies]
pysen = "^0.10.5"
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
flake8-bugbear = "^23.7.10"
mypy = "^1.5.1"
pytest = "^7.4.2"
types-requests = "^2.31.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
